корректировка шаблонов, функция цены
